Added unit tests with 95% code coverage (#2)
* Fix all detekt errors in the XMP core sources

Extract complex conditions and deeply nested blocks into named values and helper functions, replace magic numbers with named constants, split the long serializer method, and suppress LargeClass on the Adobe port classes that mirror the original structure

* Reject alias names containing path characters

Change the simple-name check in registerAlias from Regex.matches to
containsMatchIn so names like "A/B" or "B[1]" are rejected with
BADXPATH as documented, and add XMPSchemaRegistryTest covering both
names plus the valid simple-name case

* Create proper alt-text arrays in title and description setters

Set ARRAY_ALT_TEXT when creating the rdf:Alt containers in setTitle,
setDescription and the LocationName field of setLocationShown so
getLocalizedText and setLocalizedText accept them, and add
XMPMetaLocalizedTextTest covering both services for title and
description

* Fix setArrayItem KDoc to match the rejected past-the-end index

The reference Java SDK behaves like the port and throws BADINDEX for
indices beyond the array length, so document the valid range instead
of the promised appending behavior, and add XMPMetaSettersTest
guarding replace, rejection, and insert-mode appending

* Restore getOptionsString so option names are reachable

The port dropped the public getOptionsString of the Java SDK, leaving
every defineOptionName override unreachable and untestable. Restore
it with LSB-first ordering and a placeholder for undefined bits, and
cover all five option classes with OptionsGetOptionsStringTest

* Select qualifier-selector array items including the last one

Fix the off-by-one loop in lookupQualSelector so paths like
subject[?test:custom='value'] can match the final array item, and add
XMPNodeUtilsTest covering middle, last, single-item, and no-match
cases

* Skip implicit-node cleanup when none was created

Guard the deleteNode call in findNode against a null implicit root
so setProperty with [last()] on an existing empty array reports
BADXPATH instead of crashing, and cover the empty-array and
cleanup paths in XMPMetaMiscTest

* Pin rdf:datatype literal parsing behavior with tests

Probing the reported vanish-on-parse showed the value is normalized
into the canonical dc:title lang-alt array per the XMP data model and
Java SDK, so add regression tests for the datatype literal form and
the invalid element-child form

* Add comprehensive unit tests for the XMP metadata library

* Removed debug function

* Added more tests, deleted orphaned code

* Update to the latest AGP 9.3.1

* Add Kover 95% line coverage verification rule

* Add description & group to Gradle tasks

* Bumped version to 1.7.3

* Detekt failures are no longer ignored

* Fix native-target DOM access in RDF root test

Use getDocumentElement() instead of the documentElement extension
property, which is only a member on JVM/JS/Wasm and needs an import
on Kotlin/Native targets, breaking iOS and macOS CI builds

+++ src/posixMain/kotlin/de/stefan_oltmann/xmp/PosixByteArrayReader.kt
@@ -28,7 +28,12 @@ internal fun readFileAsByteArray(filePath: String): ByteArray? = memScoped {
 
     /* Move the cursor to the end of the file. */
     fseek(file, 0, SEEK_END)
-    val fileSize = ftell(file)
+
+    /*
+     * ftell returns Int on Windows and Long on Unix. Normalize to ULong so
+     * the buffer size and the expected read count share one portable type.
+     */
+    val fileSize = ftell(file).toULong()
     rewind(file)
 
     val buffer = ByteArray(fileSize.toInt())
@@ -36,13 +41,13 @@ internal fun readFileAsByteArray(filePath: String): ByteArray? = memScoped {
     val bytesReadCount: ULong = fread(
         buffer.refTo(0),
         1.toULong(), // Number of items
-        fileSize.toULong(), // Size to read
+        fileSize, // Size to read
         file
     )
 
     fclose(file)
 
-    if (bytesReadCount != fileSize.toULong()) {
+    if (bytesReadCount != fileSize) {
         perror("Did not read file completely: $bytesReadCount != $fileSize")
         return null
     }
